custom-elements/registering-custom-elements/extensions-to-document-interface/custom-element-name.html

<!DOCTYPE html>
<html>
<head>
<title>Custom element local name is the value of the EXTENDS property</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="Custom element local name is the value of the EXTENDS property">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#extensions-to-document-interface-to-register">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

function doTest(element) {

    var doc = newHTMLDocument();
    
    var name = 'x-foo-' + element;
    
    var obj = doc.createElement(element);
    
    var proto = Object.create(obj.constructor.prototype);
    
    var XFoo = doc.register(name, {
        prototype: proto,
        extends: element
    });
    
    var xfoo = new XFoo();

    assert_equals(xfoo.localName, element,
            'Custom element local name must be an EXTENDS value');

}


var parameters = [];

for (var i = 0; i < HTML5_ELEMENTS.length; i++) {
    if (HTML5_DOCUMENT_ELEMENTS.indexOf(HTML5_ELEMENTS[i]) == -1) {
        parameters[i] = ['Custom element type ' + HTML5_ELEMENTS[i], HTML5_ELEMENTS[i]];
    }
}

generate_tests(doTest, parameters, 
        'Custom element local name is the value of the EXTENDS property');


</script>
</body>
</html>
